requirements.txt
# requirements.txt
langchain>=0.1.0
langchain-community>=0.0.1
langchain-openai>=0.0.1
openai>=1.6.0
spotipy>=2.23.0
pydantic>=2.5.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
schedule>=1.2.0
# Optional dependencies
# chromadb>=0.4.18  # For vector storage if needed
# numpy>=2.1.0  # Required by some libraries
